hospital_dbms_v1-20260317T175815Z-3-001/hospital_dbms_v1/v1_final/requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
Flask-MySQLdb==2.0.0
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
mysqlclient==2.2.8
numpy==2.4.3
packaging==26.0
pillow==12.1.1
PyMySQL==1.1.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
threadpoolctl==3.6.0
urllib3==2.6.3
Werkzeug==3.1.6
